common/ceph_time: move fmt formatters to ceph_time_fmt.h
Eliminate the header dependency on fmt.
